Separate out fiscal year/qtr tests
